Refactor listgroup methods to be utility methods. Move custom methods out of aem.js into separate file
