+++ scripts/aem.js
@@ -707,60 +707,6 @@ async function waitForLCP(lcpBlocks) {
   });
 }
 
-/** JMP ADDED METHODS */
-function parseBlockOptions(block, rowName) {
-  const optionsObject = {};
-  const row = rowName === undefined ? 'options' : rowName;
-
-  const optionName = block.firstElementChild?.children.item(0).textContent;
-  if (optionName.toLowerCase() === row) {
-    const optionVal = block.firstElementChild?.children.item(1).textContent;
-    const tempOptionsArray = optionVal.length > 1 ? optionVal.split(',') : {};
-
-    tempOptionsArray.forEach((item) => {
-      if (item.includes('=')) {
-        const optionsString = item.split('=', 2);
-        optionsObject[optionsString[0]] = optionsString[1];
-      } else {
-        optionsObject[item] = true;
-      }
-    });
-  }
-  return optionsObject;
-}
-
-/**
- * Returns a list of properties listed in the block
- * @param {string} route get the Json data from the route
- * @returns {Object} the json data object
-*/
-async function getJsonFromUrl(route) {
-  try {
-    const response = await window.fetch(route);
-    if (!response.ok) return null;
-    const json = await response.json();
-    return json;
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('getJsonFromUrl:', { error });
-  }
-  return null;
-}
-
-/**
- * Returns if a given 2 or 4 digit language should have
- * a valid index. Mainly used in case of listgroup in sandbox
- * or other nonlanguage directory.
- * @param {string} language
- * @returns {Boolean} true if the index should exist.
- */
-function languageIndexExists(language) {
-  const languageIndexes = [
-    'en', 'es', 'fr', 'zh', 'de', 'it', 'ko', 'ja', 'zh-hans', 'zh-hant',
-  ];
-  return languageIndexes.includes(language);
-}
-
 init();
 
 export {
@@ -773,16 +719,13 @@ export {
   decorateSections,
   decorateTemplateAndTheme,
   fetchPlaceholders,
-  getJsonFromUrl,
   getMetadata,
-  languageIndexExists,
   loadBlock,
   loadBlocks,
   loadCSS,
   loadFooter,
   loadHeader,
   loadScript,
-  parseBlockOptions,
   readBlockConfig,
   sampleRUM,
   setup,
