+++ test/scripts/aem.test.js
@@ -5,21 +5,6 @@ import { expect } from '@esm-bundle/chai';
 const aemHelper = await import('../../scripts/aem.js');
 
 describe('aem.js tests', () => {
-  describe('Language Index', () => {
-    let languageIndexExists;
-
-    before(async () => {
-      languageIndexExists = aemHelper.languageIndexExists;
-    });
-
-    it('Language index should exist for certain languages', () => {
-      expect(true, languageIndexExists('en'));
-      expect(true, languageIndexExists('FR'));
-      expect(true, languageIndexExists('zh_hans'));
-      expect(false, languageIndexExists('en_us'));
-    });
-  });
-
   describe('Section Metadata', () => {
     let decorateSections;
